Factor out error functions into separate file
