+++ Makefile
@@ -47,7 +47,7 @@ ALL_AUTOGENERATED_SOURCES = $(AUTOGENERATED_C_FILES) $(AUTOGENERATED_H_FILES) $(
 
 LIB_BUILD_TARGET = $(SOURCE_DIR)/unicorn_c_.$(LIB_EXTENSION)
 LIB_C_SOURCES = \
-    $(addprefix $(SOURCE_DIR)/,control_functions.c engine.c hooks.c registers_api.c unicorn.c) \
+    $(addprefix $(SOURCE_DIR)/,control_functions.c engine.c errors.c hooks.c registers_api.c unicorn.c) \
     $(AUTOGENERATED_C_FILES)
 LIB_OBJECT_FILES = $(LIB_C_SOURCES:.c=.$(OBJ_EXTENSION))
