+++ csrc/unicorn.c
@@ -32,96 +32,8 @@
 #include "unicornlua/utils.h"
 #include <lauxlib.h>
 #include <lua.h>
-#include <stdarg.h>
 #include <stddef.h>
-#include <stdio.h>
 #include <unicorn/unicorn.h>
-
-void ulinternal_vsnprintf(lua_State *L, size_t max_size, const char *format, va_list argv)
-{
-    char *message = malloc(max_size + 1);
-
-    vsnprintf(message, max_size, format, argv);
-    message[max_size] = '\0';
-    lua_pushstring(L, message);
-    free(message);
-}
-
-_Noreturn int ulinternal_crash_not_implemented(lua_State *L)
-{
-    lua_pushstring(L, "BUG: This function isn't implemented in the Lua binding yet.");
-    lua_error(L);
-    UL_UNREACHABLE_MARKER;
-}
-
-void ulinternal_crash_if_failed(lua_State *L, uc_err code, const char *format, ...)
-{
-    if (code == UC_ERR_OK)
-        return;
-
-    luaL_Buffer msgbuf;
-    luaL_buffinit(L, &msgbuf);
-
-    // The general form of the produced error message is:
-    //      {user message} (Unicorn: error ### -- {unicorn message})
-
-    va_list argv;
-    va_start(argv, format);
-    ulinternal_vsnprintf(L, UL_MAX_ERROR_MESSAGE_LENGTH, format, argv);
-    va_end(argv);
-
-    luaL_addvalue(&msgbuf);
-
-    lua_pushfstring(L, " (Unicorn: error %d -- %s)", (int)code, uc_strerror(code));
-    luaL_addvalue(&msgbuf);
-
-    luaL_pushresult(&msgbuf);
-    lua_error(L);
-    UL_UNREACHABLE_MARKER;
-}
-
-_Noreturn int ulinternal_crash_unsupported_operation(lua_State *L)
-{
-    lua_pushstring(L, "The operation is not supported for this version of Unicorn.");
-    lua_error(L);
-    UL_UNREACHABLE_MARKER;
-}
-
-_Noreturn void ulinternal_crash(lua_State *L, const char *format, ...)
-{
-    va_list argv;
-    va_start(argv, format);
-    ulinternal_vsnprintf(L, UL_MAX_ERROR_MESSAGE_LENGTH, format, argv);
-    va_end(argv);
-    lua_error(L);
-    UL_UNREACHABLE_MARKER;
-}
-
-size_t count_table_elements(lua_State *L, int table_index)
-{
-    size_t count;
-
-    // Count the number of keys in the map portion of the table.
-    lua_pushnil(L);
-    for (count = 0; lua_next(L, table_index) != 0; ++count)
-        lua_pop(L, 1);
-
-        // Count the number of keys in the aray portion of the table.
-#if LUA_VERSION_NUM >= 502
-    count += (size_t)luaL_len(L, table_index);
-#else
-    for (int i = 0;; i++, count++)
-    {
-        lua_pushinteger(L, i);
-        lua_rawget(L, table_index);
-        if (lua_isnil(L, -1))
-            break;
-        lua_pop(L, 1);
-    }
-    lua_pop(L, 1);
-#endif
-    return count;
-}
 
 /**
  * Open a new Unicorn engine.
